requirements.txt
fastapi
uvicorn
supabase
mistralai
python-dotenv
pydantic
